refactor(intent): delegate tag, stuff, and plan intents to services
